Final improvements and bug fixes

+++ app/src/sharedTestUtils/java/ch/epfl/favo/FakeUserUtil.java
@@ -43,7 +43,6 @@ public class FakeUserUtil implements IUserUtil {
     return defaultResult();
   }
 
-
   public CompletableFuture defaultResult() {
     if (isThrowingError) return failedCompletableFuture;
     return successfulCompletableFuture;
@@ -82,7 +81,7 @@ public class FakeUserUtil implements IUserUtil {
     return successfulUserFuture;
   }
 
-  public CompletableFuture updateCoinBalance(String userId, double reward) {
+  public CompletableFuture updateCoinBalance(String userId, int reward) {
     return defaultResult();
   }
